Moved classes around directories a bit
