+++ tests/test_velocityDetector.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from lib.drifts.VelocityDetector import VelocityDetector
+from lib.drifts_detect.VelocityDetector import VelocityDetector
 from lib.model.Vector import Vector
